[Add] moved request struct and generator
